requirements.txt
fastapi
uvicorn
requests
boto3
python-dotenv